requirements.txt
Selenium
Appium-Python-Client
pyyaml